pyproject.toml
[project]
name = "detali-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.25.0",
    "aiomysql>=0.3.2",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "types-PyMySQL",
]
[tool.black]
line-length = 120
target-version = ["py312"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # bugbear
    "C4",   # comprehensions
    "SIM",  # simplifications
]
ignore = ["SIM105", "SIM108", "SIM401"]  # try/except; if/else; row.get (sqlite3.Row в стабах без .get)

[tool.ruff.lint.isort]
known-first-party = ["bot"]

# Декораторы aiogram и наследование от BaseModel/StatesGroup без полных стабов
[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_unused_ignores = true
strict = true
disallow_any_generics = true
no_implicit_optional = true
ignore_missing_imports = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["bot.handlers.*", "bot.middlewares.*", "bot.main"]
disallow_untyped_defs = false
disallow_subclassing_any = false
disable_error_code = ["misc", "attr-defined"]
[[tool.mypy.overrides]]
module = ["bot.database.models"]
disallow_subclassing_any = false
[[tool.mypy.overrides]]
module = ["bot.services.catalog_sync"]
disable_error_code = ["import-untyped"]
[[tool.mypy.overrides]]
module = ["bot.core.opencart_client", "bot.handlers.account"]
disable_error_code = ["no-any-return"]
